Make --archive-dir in remove-older-than match
When `archive_directory` and `remove_older_than` parameters are provided
then both commands in the spool file should have the same `--archive-dir`
argument.

Without this we'll be storing some cache files in two separate places. It
also seems to be the cause of Duplicity continuing to need the secret key
after we accidentally deleted our cache directory. Testing it with the same
argument for both commands stopped it prompting.