Add multi-processing and part chunk size opts
This adds the two duplicity options:

```
--s3-use-multiprocessing
--s3-multipart-chunk-size=number
```

When multiprocessing is enabled, it will upload files to S3 using
multiple processes by chunking files into parts. The number of processes
is defaulted to the number of detected CPUs on the system.

The chunk size number is defaulted to 25MB, but this can be changed to a
larger or smaller size depending on the requirement. Smaller chunks can
be to retrieve for a restore, but it will create a larger number of
files.

The maximum chunk size of S3 is 5GB.